Card design corrections. Teacher build fix.
