+++ libs/pandautils/src/main/java/com/instructure/pandautils/di/HomeroomViewModelModule.kt
@@ -1,0 +1,22 @@
+package com.instructure.pandautils.di
+
+import android.content.res.Resources
+import com.instructure.canvasapi2.managers.AnnouncementManager
+import com.instructure.canvasapi2.managers.PlannerManager
+import com.instructure.canvasapi2.managers.UserManager
+import com.instructure.pandautils.features.elementary.homeroom.CourseCardCreator
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+
+@Module
+@InstallIn(ViewModelComponent::class)
+class HomeroomViewModelModule {
+
+    @Provides
+    fun provideCourseCardCreator(plannerManager: PlannerManager, userManager: UserManager,
+                                 announcementManager: AnnouncementManager, resources: Resources): CourseCardCreator {
+        return CourseCardCreator(plannerManager, userManager, announcementManager, resources)
+    }
+}
